RATEST-268: Refactoring TestData class and PatientGenerator class into…
