+++ qaframework-bdd-tests/src/main/java/org/openmrs/contrib/qaframework/helper/TestData.java
@@ -153,9 +153,8 @@ public class TestData {
 
 	/**
 	 * Add a Person to the database and return its uuid.
-	 * 
 	 * @param person
-	 *            The (test) Person to add to the database.
+	 * The (test) Person to add to the database.
 	 * @return The new Person's uuid.
 	 */
 	public static String createPerson(PersonInfo person) {
@@ -170,9 +169,8 @@ public class TestData {
 
 	/**
 	 * Create a location without tags
-	 * 
 	 * @param locationName
-	 *            the name of the location
+	 * the name of the location
 	 * @return the location's uuid
 	 */
 	public static String createLocation(String locationName) {
@@ -183,9 +181,8 @@ public class TestData {
 
 	/**
 	 * Delete a resource permanently
-	 * 
 	 * @param uuid
-	 *            of the resource
+	 * of the resource
 	 */
 	public static void permanentDelete(String uuid) {
 
@@ -241,7 +238,6 @@ public class TestData {
 
 	/**
 	 * Check if patient with given id exists
-	 * 
 	 * @param id
 	 * @return true if patient exists, false otherwise
 	 */
@@ -275,14 +271,12 @@ public class TestData {
 
 		/**
 		 * Return the REST name of the object/class.
-		 * 
 		 * @return the REST name of the object/class.
 		 */
 		public abstract String name();
 
 		/**
 		 * Create the object in the database using REST. Return the uuid.
-		 * 
 		 * @return uuid of the new object.
 		 */
 		public String create() {
@@ -684,4 +678,61 @@ public class TestData {
 		public String datetime;
 		public String uuid;
 	}
+
+	public static class PatientGenerator {
+
+		private static final String[] PATIENT_GIVEN_NAMES = {"Alexandre", "Achint",
+				"Brittany", "Burke", "Cesar", "Cosmin", "Daniel", "Darius",
+				"David", "Ellen", "Émerson", "Evan", "Fernando", "Gabou",
+				"Glauber", "Hamish", "Louise", "Mário", "Mark", "Natália", "Neil",
+				"Neissi", "Nelice", "Rafal", "Renee", "Wyclif"};
+		private static final String[] PATIENT_FAMILY_NAMES = {"Barbosa", "Sethi",
+				"Eddy", "Mamlin", "Vortmann", "Ioan", "Kayiwa", "Jazayeri",
+				"Walton", "Ball", "Hernandez", "Waters", "Freire", "Mandy",
+				"Ramos", "Fraser", "Sécordel", "Areias", "Goodrich", "Arsand",
+				"Craven", "Lima", "Heck", "Korytkowski", "Orser", "Luyima"};
+		private static final String[] PATIENT_BIRTH_MONTH = MONTHS;
+		private static final String[] PATIENT_BIRTH_DAY = DAYS;
+		private static final String[] PATIENT_BIRTH_YEAR = YEARS;
+		private static final String[] PATIENT_GENDER = {"Male", "Female"};
+
+		public static org.openmrs.contrib.qaframework.helper.TestPatient generateTestPatient() {
+			org.openmrs.contrib.qaframework.helper.TestPatient p = new org.openmrs.contrib.qaframework.helper.TestPatient();
+			p.givenName = randomArrayEntry(PATIENT_GIVEN_NAMES);
+			p.middleName = "";
+			p.familyName = randomArrayEntry(PATIENT_FAMILY_NAMES);
+			p.gender = randomArrayEntry(PATIENT_GENDER);
+			p.birthDay = randomArrayEntry(PATIENT_BIRTH_DAY);
+			p.birthMonth = randomArrayEntry(PATIENT_BIRTH_MONTH);
+			p.birthYear = randomArrayEntry(PATIENT_BIRTH_YEAR);
+			String suffix = randomSuffix();
+			p.address1 = "Address1" + suffix;
+			p.address2 = "Address2" + suffix;
+			p.city = "City" + suffix;
+			p.state = "State" + suffix; // TODO shorter string for State perhaps?
+			p.country = "Country" + suffix; // TODO shorter string for Country
+			// perhaps?
+			p.phone = randomSuffix(9);
+			p.postalCode = "345234"; // hardwired for now
+			p.latitude = "12"; // hardwired for now
+			p.longitude = "47"; // hardwired for now
+			p.startDate = "01-01-2000"; // hardwired for now
+			p.endDate = "01-01-2010"; // hardwired for now
+			return p;
+		}
+
+		static String randomArrayEntry(String[] array) {
+			return array[(int) (Math.random() * array.length)];
+		}
+
+		static String randomSuffix() {
+			return randomSuffix(6);
+		}
+
+		static String randomSuffix(int digits) {
+			// First n digits of the current time.
+			return String.valueOf(System.currentTimeMillis()).substring(0, digits - 1);
+		}
+	}
+
 }
