+++ qaframework-legacy-tests/src/test/java/org/openmrs/contrib/qaframework/legacy/DuplicatePatientRegisterTest.java
@@ -24,7 +24,6 @@ import org.junit.experimental.categories.Category;
 import org.openmrs.contrib.qaframework.helper.BuildTests;
 import org.openmrs.contrib.qaframework.helper.TestData;
 import org.openmrs.contrib.qaframework.helper.TestPatient;
-import org.openmrs.contrib.qaframework.helper.PatientGenerator;
 import org.openmrs.contrib.qaframework.page.ClinicianFacingPatientDashboardPage;
 import org.openmrs.contrib.qaframework.page.RegistrationPage;
 
@@ -34,7 +33,7 @@ public class DuplicatePatientRegisterTest  extends ReferenceApplicationTestBase 
 
     @Before
     public void setUp() throws AWTException {
-        patient = PatientGenerator.generateTestPatient();
+        patient = TestData.PatientGenerator.generateTestPatient();
     }
 
     @Test
